requirements_torchrl.txt
# TorchRL Training Dependencies
torch>=2.1.0
torchvision>=0.16.0
torchrl>=0.3.0
tensordict>=0.3.0
gymnasium>=0.29.0
pillow>=10.0.0
numpy>=1.24.0
hydra-core>=1.3.0
playwright>=1.40.0
wandb  # Optional for logging
